# app/Http/Controllers/Api/AdminController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Http\JsonResponse;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Facades\Validator;

/**
 * @OA\Tag(
 *     name="Admin",
 *     description="Admin management endpoints"
 * )
 */
class AdminController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/admin/stats",
     *     summary="Get admin dashboard statistics",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Admin statistics retrieved successfully"
     *     ),
     *     @OA\Response(response=403, description="Forbidden - Admin access required")
     * )
     */
    public function getStatistics(): JsonResponse
    {
        try {
            // Get various statistics
            $totalUsers = User::count('*');
            $totalMitra = User::where('role', ' =>', 'mitra', 'and')->count('*');
            $totalOrders = DB::table('orders')->count('*');
            $totalServices = DB::table('services')->count('*');
            $pendingOrders = DB::table('orders')->where('status', ' =>', 'pending', 'and')->count('*');
            $completedOrders = DB::table('orders')->where('status', ' =>', 'completed', 'and')->count('*');
            $activeUsers = User::where('status', ' =>', 'active', 'and')->count('*');
            $activeMitra = User::where('role', ' =>', 'mitra', 'and')->where('status', ' =>', 'active', 'and')->count('*');

            // Revenue calculations
            $totalRevenue = DB::table('payments')->where('status', ' =>', 'paid', 'and')->sum('amount');
            $monthlyRevenue = DB::table('payments')
                ->where('status', ' =>', 'paid', 'and')
                ->whereMonth('created_at', ' =>', now()->month, 'and')
                ->sum('amount');

            // Orders by status
            $ordersByStatus = DB::table('orders')
                ->select('status', DB::raw('count(*) as count'))
                ->groupBy('status')
                ->pluck('count', 'status')
                ->toArray();

            // Users by role
            $usersByRole = User::select('role', DB::raw('count(*) as count'))
                ->groupBy('role')
                ->pluck('count', 'role')
                ->toArray();

            // Recent activities (simplified)
            $recentActivities = [
                [
                    'type' => 'order',
                    'description' => 'New order created',
                    'timestamp' => now()->subMinutes(5),
                ],
                [
                    'type' => 'user',
                    'description' => 'New user registered',
                    'timestamp' => now()->subMinutes(15),
                ],
            ];

            return response()->json([
                'success' => true,
                'data' => [
                    'total_users' => $totalUsers,
                    'total_mitra' => $totalMitra,
                    'total_orders' => $totalOrders,
                    'total_services' => $totalServices,
                    'pending_orders' => $pendingOrders,
                    'completed_orders' => $completedOrders,
                    'active_users' => $activeUsers,
                    'active_mitra' => $activeMitra,
                    'total_revenue' => $totalRevenue,
                    'monthly_revenue' => $monthlyRevenue,
                    'orders_by_status' => $ordersByStatus,
                    'users_by_role' => $usersByRole,
                    'recent_activities' => $recentActivities,
                ]
            ]);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to get admin statistics',
                'error' => $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/admin/users",
     *     summary="Get all users for admin management",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Parameter(name="page", in="query", description="Page number", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="limit", in="query", description="Items per page", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="role", in="query", description="Filter by role", @OA\Schema(type="string")),
     *     @OA\Parameter(name="status", in="query", description="Filter by status", @OA\Schema(type="string")),
     *     @OA\Parameter(name="search", in="query", description="Search term", @OA\Schema(type="string")),
     *     @OA\Response(response=200, description="Users retrieved successfully")
     * )
     */
    public function getUsers(Request $request): JsonResponse
    {
        try {
            $query = User::query();
            
            // Apply filters
            if ($request->has('role')) {
                $query->where('role', ' =>', $request->role, 'and');
            }
            
            if ($request->has('status')) {
                $query->where('status', ' =>', $request->status, 'and');
            }
            
            if ($request->has('search')) {
                $search = $request->search;
                $query->where(function($q) use ($search) {
                                            $q->where('name', 'like', "%{$search}%")
                                                ->orWhere('email', 'like', "%{$search}%");
                });
            }
            
            $users = $query->paginate($request->get('limit', 10));
            
            return response()->json([
                'success' => true,
                'data' => $users
            ]);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to get users',
                'error' => $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Post(
     *     path="/api/admin/users",
     *     summary="Create new admin user",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"name","email","password","role"},
     *             @OA\Property(property="name", type="string"),
     *             @OA\Property(property="email", type="string"),
     *             @OA\Property(property="password", type="string"),
     *             @OA\Property(property="role", type="string")
     *         )
     *     ),
     *     @OA\Response(response=201, description="User created successfully")
     * )
     */
    public function createUser(Request $request): JsonResponse
    {
        $validator = Validator::make($request->all(), [
            'name' => 'required|string|max:255',
            'email' => 'required|string|email|max:255|unique:users',
            'password' => 'required|string|min:8',
            'role' => 'required|string|in:admin,mitra,end_user',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        try {
            $user = User::create([
                'name' => $request->name,
                'email' => $request->email,
                'password' => Hash::make($request->password),
                'role' => $request->role,
                'status' => 'active',
                'email_verified_at' => now(),
            ]);

            return response()->json([
                'success' => true,
                'message' => 'User created successfully',
                'data' => $user
            ], 201);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to create user',
                'error' => $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Patch(
     *     path="/api/admin/users/{id}",
     *     summary="Update user status",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="integer")),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string")
     *         )
     *     ),
     *     @OA\Response(response=200, description="User updated successfully")
     * )
     */
    public function updateUser(Request $request, $id): JsonResponse
    {
        $validator = Validator::make($request->all(), [
            'status' => 'sometimes|string|in:active,blocked,pending',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        try {
            $user = User::findOrFail($id);
            $user->update($request->only(['status']));

            return response()->json([
                'success' => true,
                'message' => 'User updated successfully',
                'data' => $user
            ]);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to update user',
                'error' => $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Delete(
     *     path="/api/admin/users/{id}",
     *     summary="Delete user",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="integer")),
     *     @OA\Response(response=200, description="User deleted successfully")
     * )
     */
    public function deleteUser($id): JsonResponse
    {
        try {
            $user = User::findOrFail($id);
            $user->delete();

            return response()->json([
                'success' => true,
                'message' => 'User deleted successfully'
            ]);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to delete user',
                'error' => $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/admin/logs",
     *     summary="Get system logs",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Response(response=200, description="Logs retrieved successfully")
     * )
     */
    public function getLogs(Request $request): JsonResponse
    {
        // Simplified log implementation
        $logs = [
            [
                'level' => 'info',
                'message' => 'User login successful',
                'timestamp' => now()->subMinutes(5),
                'context' => ['user_id' => 1]
            ],
            [
                'level' => 'warning',
                'message' => 'Failed login attempt',
                'timestamp' => now()->subMinutes(10),
                'context' => ['email' => 'test@example.com']
            ]
        ];

        return response()->json([
            'success' => true,
            'data' => [
                'data' => $logs,
                'current_page' => 1,
                'last_page' => 1,
                'per_page' => 20,
                'total' => count($logs)
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/admin/export",
     *     summary="Export data",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Response(response=200, description="Export URL generated")
     * )
     */
    public function exportData(Request $request): JsonResponse
    {
        // Simplified export implementation
        return response()->json([
            'success' => true,
            'data' => [
                'download_url' => '/exports/data_' . time() . '.csv'
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/admin/notifications",
     *     summary="Send system notification",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Response(response=200, description="Notification sent successfully")
     * )
     */
    public function sendNotification(Request $request): JsonResponse
    {
        $validator = Validator::make($request->all(), [
            'title' => 'required|string|max:255',
            'message' => 'required|string',
            'role' => 'sometimes|string|in:admin,mitra,end_user',
            'user_ids' => 'sometimes|array',
            'user_ids.*' => 'integer|exists:users,id'
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        // Simplified notification implementation
        return response()->json([
            'success' => true,
            'message' => 'Notification sent successfully'
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/admin/health",
     *     summary="Get system health",
     *     tags={"Admin"},
     *     security={{"bearer_token":{}}},
     *     @OA\Response(response=200, description="System health retrieved")
     * )
     */
    public function getSystemHealth(): JsonResponse
    {
        return response()->json([
            'success' => true,
            'data' => [
                'database' => 'ok',
                'redis' => 'ok',
                'storage' => 'ok',
                'memory_usage' => '50%',
                'disk_usage' => '30%',
                'uptime' => '15 days'
            ]
        ]);
    }
}