# public/openapi.yaml
openapi: 3.0.3
info:
  title: Gerobaks API
  description: |
    # 🚛 Gerobaks Waste Management API
    
    RESTful API Backend untuk sistem manajemen pengumpulan sampah Gerobaks.
    
    ## 🌟 Features
    - 🔐 Authentication & Authorization (Laravel Sanctum)
    - 📅 Schedule Management (Mobile & Standard Format)
    - 📍 Real-time GPS Tracking
    - 🛒 Order Management
    - 💰 Payment & Balance System
    - ⭐ Rating & Reviews
    - 💬 Chat System
    - 🔔 Notifications
    - 📊 Dashboard & Analytics
    - 📝 Reports & Feedback
    - 🎯 Subscription Plans
    - ⚙️ Settings Management
    - 👨‍💼 Admin Management
    
    ## 📖 Repository
    - **GitHub**: [https://github.com/fk0u/gerobackend](https://github.com/fk0u/gerobackend)
    - **Developer**: [@fk0u](https://github.com/fk0u)
    
    ## 🔐 Authentication
    API menggunakan **Laravel Sanctum** untuk autentikasi.
    
    **Flow:**
    1. POST /api/login → Get token
    2. Use token in header: `Authorization: Bearer {token}`
    3. POST /api/auth/logout → Revoke token
    
  version: 1.0.0
  contact:
    name: fk0u
    url: https://github.com/fk0u
    email: support@gerobaks.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gerobaks.dumeg.com/api
    description: Production Server
  - url: https://staging-gerobaks.dumeg.com/api
    description: Staging Server
  - url: http://127.0.0.1:8000/api
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Login untuk mendapatkan token:
        ```
        Authorization: Bearer YOUR_TOKEN_HERE
        ```
...
paths:
...
  /admin/stats:
    get:
      tags: [Admin]
      summary: Get System Statistics
      description: Admin only
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  total_mitra:
                    type: integer
                  total_schedules:
                    type: integer
                  total_revenue:
                    type: number
                  active_schedules:
                    type: integer
  /admin/users:
    get:
      tags: [Admin]
      summary: Get All Users
      description: Admin only
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users list

    post:
      tags: [Admin]
      summary: Create User
      description: Admin only
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User created
  /admin/users/{id}:
    patch:
      tags: [Admin]
      summary: Update User
      description: Admin only
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User updated

    delete:
      tags: [Admin]
      summary: Delete User
      description: Admin only
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
  /admin/logs:
    get:
      tags: [Admin]
      summary: Get System Logs
      description: Admin only
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System logs
  /admin/export:
    get:
      tags: [Admin]
      summary: Export Data
      description: Admin only
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [users, schedules, orders, payments]
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, excel, pdf]
      responses:
        '200':
          description: Export file
  /admin/notifications:
    post:
      tags: [Admin]
      summary: Send Broadcast Notification
      description: Admin only
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, message]
              properties:
                title:
                  type: string
                message:
                  type: string
                target:
                  type: string
                  enum: [all, end_user, mitra]
      responses:
        '201':
          description: Notification sent
...